requirements.txt
flask==2.0.1
tensorflow==2.7.0
pillow==9.0.0
numpy==1.21.2
python-dotenv==0.19.0
openai==0.27.0fl